feat(sw): :recycle: Refactoring to TS
